Switch from int to long for equation result and operands
needed because numbers get very big
adapt functional and test code

+++ day_07/java/app/src/main/java/tkrippes/com/github/adventofcode2024/day07/Equation.java
@@ -2,5 +2,5 @@ package tkrippes.com.github.adventofcode2024.day07;
 
 import java.util.List;
 
-public record Equation(int result, List<Integer> operands) {
+public record Equation(long result, List<Long> operands) {
 }
